requirements.txt
Flask==3.0.*
PyYAML==6.*
requests==2.*
pytest